fix CodeQL v1

+++ hugegraph-test/src/main/java/com/baidu/hugegraph/core/EdgeCoreTest.java
@@ -5218,7 +5218,7 @@ public class EdgeCoreTest extends BaseCoreTest {
         while (page != null) {
             GraphTraversal<?, ?> iterator = graph.traversal().V(james).bothE()
                                                  .has("~page", page).limit(1);
-            long size = IteratorUtils.count(iterator);
+            Long size = IteratorUtils.count(iterator);
             if (size == 0L) {
                 // The last page is empty
                 Assert.assertEquals(6, count);
@@ -5226,7 +5226,7 @@ public class EdgeCoreTest extends BaseCoreTest {
                 Assert.assertEquals(1, size);
             }
             page = TraversalUtil.page(iterator);
-            count += size;
+            count += size.intValue();
         }
         Assert.assertEquals(6, count);
     }
@@ -5244,7 +5244,7 @@ public class EdgeCoreTest extends BaseCoreTest {
         while (page != null) {
             GraphTraversal<?, ?> iterator = graph.traversal().V(james).outE()
                                                  .has("~page", page).limit(1);
-            long size = IteratorUtils.count(iterator);
+            Long size = IteratorUtils.count(iterator);
             if (size == 0L) {
                 // The last page is empty
                 Assert.assertEquals(4, count);
@@ -5252,7 +5252,7 @@ public class EdgeCoreTest extends BaseCoreTest {
                 Assert.assertEquals(1, size);
             }
             page = TraversalUtil.page(iterator);
-            count += size;
+            count += size.intValue();
         }
         Assert.assertEquals(4, count);
     }
@@ -5270,7 +5270,7 @@ public class EdgeCoreTest extends BaseCoreTest {
         while (page != null) {
             GraphTraversal<?, ?> iterator = graph.traversal().V(james).inE()
                                                  .has("~page", page).limit(1);
-            long size = IteratorUtils.count(iterator);
+            Long size = IteratorUtils.count(iterator);
             if (size == 0L) {
                 // The last page is empty
                 Assert.assertEquals(2, count);
@@ -5278,7 +5278,7 @@ public class EdgeCoreTest extends BaseCoreTest {
                 Assert.assertEquals(1, size);
             }
             page = TraversalUtil.page(iterator);
-            count += size;
+            count += size.intValue();
         }
         Assert.assertEquals(2, count);
     }
@@ -7516,10 +7516,10 @@ public class EdgeCoreTest extends BaseCoreTest {
         int count = 0;
         while (page != null) {
             GraphTraversal<?, ?> iterator = fetcher.apply(page);
-            long size = IteratorUtils.count(iterator);
+            Long size = IteratorUtils.count(iterator);
             Assert.assertLte(1L, size);
             page = TraversalUtil.page(iterator);
-            count += size;
+            count += size.intValue();
         }
         return count;
     }
